level refactor
updated collision
