+++ examples/roots_of_darknes/src/level.hpp
@@ -3,35 +3,34 @@
 
 #include <elemd/vec2.hpp>
 #include <map>
-#include <string>
+#include <vector>
 
 #define MAP_TILE_SIZE 16
 #define SPRITE_SCALAT_POT 1
 
 class Player;
 
-struct Block
+struct MapTile
 {
-    unsigned char id;
-    bool collide;
-    bool render;
-    elemd::vec2 uv;
+    unsigned char tileSymbol;
+    int offsetX;
+    int offsetY;
+    int row;
+    int col;
 };
 
 class Level
 {
 public:
-    static std::map<unsigned int, Block> BlockDb;
+    std::vector<MapTile> backgroundLayer;
+    std::vector<MapTile> collisionLayer;
+    std::vector<int> collisionIndex;
 
-    
-    ~Level();
-    Block getAt(int x, int y);
     bool getCollideAtWorld(float x, float y);
     int getTileSize();
     int getTextureSize();
     int getRows();
     int getCols();
-    const unsigned char* getData();
     void loadLevelFile(std::string filePath, Player* p);
 
 private:
@@ -39,8 +38,6 @@ private:
     int _textureSize = MAP_TILE_SIZE * SPRITE_SCALAT_POT;
     int _rows = -1;
     int _cols = -1;
-
-    unsigned char* _data = nullptr;
 };
 
 #endif // LEVEL_HPP
